fix: DTDs with 165MHz+ clock ignored on displays with max_tmds_clock
In 976e96bbc07a499e009965b8defab255c1042ef0, I assumed that the version
we are getting in edid is HDMI version. But that assumption was wrong.
Because filtering was done based on the version and not based on whether
tmds clock was present or not, the higher resolutions even on compliant
monitors gets ignored and we end up booting with less resolution.

This commit fixes the same.

Change-Id: I840a6a80127c1c38d92d069b5fd15224d2917ff0